requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Firebase Admin SDK
firebase-admin==6.2.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and Storage
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Math and OCR Processing
sympy==1.12
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3

# LLM Integration
openai==1.3.8

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
python-dotenv==1.0.0
